cannon: Pull loading functionality out of patch.go
